Exercise 1 - Complete
Decisions:
1. Reduce the amount of components.
After I have read the assignments more in-depth I found out that is
better instead of having three separated components to have one that is
building over the following exercise.
2. Create box and container in order to centre the element. The picture
suggest there must be a input field with an search icon. I assume that
there will desktop and tablet version so I have created extra media
query to scale correctly.
3. Search icon. I only have created the placeholder for the icon(span).
We can use different methods for creating the icon like using a Font
Awesome pack, an image or SVG image. However, for the time
consideration I don’t think it is important to be added. I used only
text to mark it